backend/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
appdirs==1.4.4
asgiref==3.3.1
astroid==2.5
autopep8==1.5.5
colorama==0.4.4
distlib==0.3.1
filelock==3.0.12
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
psycopg2==2.8.6
pycodestyle==2.6.0
pylint==2.7.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
virtualenv==20.4.2
wrapt==1.12.1
django-environ==0.4.5
djangorestframework==3.12.0
PyJWT~=1.7.1
djangorestframework-simplejwt=4.6.0